Added a command_exists function to make that simpler.
The traditional method requires redirection and such which is always
hard to remember.  This function just returns 1 and 0 and can be treated
as a simple boolean result.

Fixed the path script and the setup script to use the new function.  For
the setup script I am first verifying that both rcm and curl are
installed.