Now recognizing WSL (windows) as a distro.
Linux will still be the "primary" os but the secondary or distro will be
recognized as "windows".  Thus the windows scripts need to move to the
distro folder.  This sets up the ability to once again support Git For
Windows or other things like that (not sure I will, but still).