Changing detection for distros and WSL (windows).
I'm splitting out the distinction of OS and whether I'm running in WSL.
WSL will now be treated as a special case with scripts that run last
over any os and distro-specific scripts.  Also fixed the distro
detection, which broke when I was trying to incorporate WSL into the
detection process.  Now using the fairly standard /etc/os-release to
determine the Linux distribution.