Changing vim cut\paste bindings, removing paste
Removed the paste.exe since it wasn't working from within the WSL shell.
It needs to be copied into the system which I'll just do from now on
during a machine setup.

Also, switched from vim-easyclip, to vim-cutlass which contains just the
parts of vim-easyclip that I really wanted (mainly the black-hole for
traditional cut commands).