Use wslpath to path manipulation of diffs
Previously I was using sed and in-line string manipulation to convert
slashes to work on windows.  WSL now provides wslpath to take care of
that for you.  So, switched to using it.

Still can't visual diff files that are in the unix path, but that is
coming in the next Windows 10.