Changing setup.sh to not call mkrc
I'm changing this for two reasons.  First, is that there is no way to
suppress running the pre and post hooks on mkrc and when setting things
up the prerequisites for those may not be ready.  Second, I don't
necessarily want to keep track of the .rcrc file of every machine I set
up.