fix: Better fix for xdg-open use in WSL
Ok, lots of trial and error with different attempts on making xdg-open
work as smoothly as possible under WSL.  Extra bonus is that I'm not
replacing xdg-open like the old version, which did produce some
non-working edge cases.  This version uses a shell script called
wsl-open as the main way to do it.  Associations do need to be created
in order for everything to work and so WSL now has a .mailcap file in
the rcs that gets placed in the home dir.