feat: Adding base-profile.bash, updating setup
The base-profile script will now serve as a core script to be called
from any script as well as automatically loaded into my bash profile for
normal shells.  This script defines and exports environment variables
for key path locations.  It integrates with the XDG Directory
Specification but does not require its use.  The script also creates
(but does not export) a host of console text color and effects variables
to be used by scripts.  Updated the initial setup script to use this
new base-profile script.