fix(asdf): No longer using 'lts' for node
The "redirects" in asdf are a nice idea but they don't really function.
If you update a plugin that then points the 'lts' to something else it
stops working in your environment and yet asdf will not install the
newer version for you.  So, in the end it causes more pain than it is
worth.  Instead, just manually update the version whenever is
convenient.