feat: Replace resetHeadset with resetUsb
Doing the headset on its own wasn't solving the problem.  The USB hubs
attached are part of the problem.  Therefore, it is best/easier to
simply reset all USB devices.  This is a rather simplistic version, by
just going in order.  It gets a mix of hubs and devices, when really we
should do all hubs first, then all devices.  I may work on a more
sophisticated version another time.