feat: Converting yt utility from session sourced to script
I still have a few bash "functions" that are expressed as sourced scripts that embed themselves into
the shell. It is more efficient to instead call those as a script when needed (and easier, since I
can edit them without have to reload the shell). This step converts one of my most used utilities
to a script rather than shell sourced. The wrapper utilities for yt-dlp.