feat: Improved yt format selection, separated list download
The yt command will now only download the target video rather than the whole playlist if a combined
url is sent in. Instead, use the ytlist to download a playlist. However, passing only a playlist
url to either command will download the list. Also improved format selection for the video.