feat: yt utilities now support passing url
The yt commands will now support passing in the url rather than requiring the use
of the clipboard. The url must be the FIRST parameter passed in as all other parameters
are passed through to the yt-dlp command. If the URL is not passed in the commands will
fall back to their current behavior and use the url from the clipboard.