fix(scripts): The resetusb script now prompts for sudo
Rather than having to type sudo to run the script the script is now internally
prompting for sudo when needed.  This is better as I can call resetusb, which is on the path, as
opposed to having to put the path in when I use sudo at the command line.