feat(colorscript): My custom implementation of shell-color-scripts
See https://gitlab.com/dwt1/shell-color-scripts for original script and concept. I simplified it a fair bit but added
support for having "sets" of files, so they can be better targeted for their purpose. For instance, I created a
'banners' set which are scripts ideal for a terminal banner or as banners in NeoVim.