feat(nvim): Begin of nvim plugin reconfig, streamline
Starting to significantly adjust plugins being used. Adding in mini and snacks and using those two to replace a number
of other plugins. This commit removes alpha as the dashboard, the indentline plugin, and lays the foundation for
removing a bunch of others. Also have plans to convert to native LSP configuration.