fix(moosefs): Finally fixed long start time when moose is down
Needed to add a connectivity check to see if the moosefs server is even alive before we check the mount point. If mount
never happened it was no big deal, but if mount previously happened and in the interim the server died we get a long 2
minute timeout (from fuse) before it fails to stat the mount point.  Doing a tcp connectivity check is much quicker.

Might consider forcing an unmount if the server is offline (making bashrc self-heal both server loss and server return).