fix(history): Another attempt at a history clean function
Trying different ideas to find the one I like best.  At minimum I want the commands from the current session to be
recorded in history (and deduped) when a bash session terminates.  Even better would be to have other sessions all sync
but somehow keep the recency of the current session at the bottom (most recent in history).