fix(nvim): No longer setting color scheme in plugin folder
Now we are setting the background and theme directly in init.lua. This will allow multiple themes to be installed and is
setting up for a possible future where I can change themes on-the-fly. Before theme switching can be made to work well,
I first need to be able to toggle between light and dark (which affects more than just Neovim, it effects the terminal
and Tmux). Worst case I may be in a situation where to toggle light/dark may require the closing and reloading of all
terminals, best case scenario I figure out a way to change it dynamically.