refactor(nvim): Split tools files into data and just the utils library
The utils library is the only thing with util methods, the other files are all data stores or read only sets of config
and\or info. A data folder makes more sense for those and I'm moving utils to root of lua folder so it can be loaded
with the simpler require("utils").