fix(aliases): Fixed classify option on ls so it can more easily be overridden
I added it to LS_OPTIONS in a way that made it impossible to override. Now it is added in a way that can be easily
overridden and also doesn't show for short lists (which is preferred there given the usual goal for that listing - copy
the list of file names). Also added new alias 'catp' to run cat/bat without paging.